requirements.txt
Django>=5.1
djangorestframework
python-dotenv
cryptography
PyJWT
requests
django-cors-headers
gunicorn
django-redis
redis
whitenoise
cachetools